Initial splitting and re-arranging to support Serverless check types
