+++ checkov/serverless/__init__.py
@@ -1,1 +1,1 @@
-from checkov.serverless.checks.aws import *
+from checkov.serverless.checks.function import *
