+++ checkov/serverless/base_function_check.py
@@ -1,0 +1,4 @@
+# NOTE: Leaving this around for backwards compatibility. Usage of the checks/function/base_function_check.py
+#       version is preferred
+# noinspection PyUnresolvedReferences
+from checkov.serverless.checks.function.base_function_check import BaseFunctionCheck
