+++ checkov/serverless/checks/function/__init__.py
@@ -1,0 +1,1 @@
+from checkov.serverless.checks.function.aws import *
