+++ checkov/serverless/checks/function/base_registry.py
@@ -1,0 +1,7 @@
+from checkov.serverless.base_registry import BaseServerlessRegistry
+
+
+class Registry(BaseServerlessRegistry):
+
+    def __init__(self):
+        super().__init__("function")
