+++ checkov/serverless/checks/function/registry.py
@@ -1,0 +1,3 @@
+from checkov.serverless.checks.function.base_registry import Registry
+
+function_registry = Registry()
