+++ tests/serverless/checks/aws/test_StarActionPolicyDocument.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from checkov.serverless.checks.aws.StarActionPolicyDocument import check
+from checkov.serverless.checks.function.aws.StarActionPolicyDocument import check
 from checkov.serverless.runner import Runner
 from checkov.runner_filter import RunnerFilter
